requirements.txt
python-telegram-bot==21.4
pyaml==24.7.0
python-telegram-bot[job-queue]